+++ R/das_format_pdf.R
@@ -17,9 +17,7 @@
 #'   \code{TRUE} if writing of file was successful, and \code{FALSE} otherwise
 #'
 #' @examples
-#' \donttest{
-#' # das_format_pdf("DAS_Format.pdf", overwrite = FALSE)
-#' }
+#' if (interactive()) das_format_pdf("DAS_Format.pdf", overwrite = FALSE)
 #'
 #' @export
 das_format_pdf <- function(file, ...) {