+++ R/swfscDAS-internal.R
@@ -5,6 +5,7 @@
 
 
 ###############################################################################
+# Copied to swfscAirDAS - update in both places
 # Internal for printing more useful error/warning message with file/line_num
 .print_file_line <- function(file.das, line.num, print.which) {
   ###Inputs