+++ data-raw/das_sample_stratum.R
@@ -1,8 +1,9 @@
 ## code to prepare `das_sample_stratum` dataset goes here
+# Note will need to update this is a new das_sample file is created because of random numbers
 
 df.out <- data.frame(
-  Lon = c(-142, -142, -138, -138, -142),
-  Lat = c(42, 44.5, 44.5, 42, 42)
+  Lon = c(-138, -138, -136.5, -136.5, -138),
+  Lat = c(39, 40.7, 40.7, 39, 39)
 )
 
 write.csv(df.out, file = "inst/das_sample_stratum.csv", row.names = FALSE)